+++ internal/backoff/backoff_test.go
@@ -9,6 +9,13 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+// exp represents a struct containing the lower and upper bounds of a time duration.
+type exp struct {
+	eq  time.Duration
+	gte time.Duration
+	lte time.Duration
+}
+
 func TestDelays(t *testing.T) {
 	duration := func(s string) (d time.Duration) {
 		d, err := time.ParseDuration(s)
@@ -42,11 +49,6 @@ func TestDelays(t *testing.T) {
 }
 
 func TestLogBackoff(t *testing.T) {
-	type exp struct {
-		eq  time.Duration
-		gte time.Duration
-		lte time.Duration
-	}
 	for _, tt := range []struct {
 		backoff Backoff
 		exp     []exp
@@ -135,31 +137,46 @@ func TestLogBackoff(t *testing.T) {
 			for seed := int64(0); seed < tt.seeds; seed++ {
 				for n, exp := range tt.exp {
 					act := tt.backoff.Delay(n)
-					if eq := exp.eq; eq != 0 {
-						if eq != act {
-							t.Fatalf(
-								"unexpected Backoff delay: %s; want %s",
-								act, eq,
-							)
-						}
-
-						continue
-					}
-					if gte := exp.gte; act <= gte {
-						t.Errorf(
-							"unexpected Backoff delay: %s; want >= %s",
-							act, gte,
-						)
-					}
-					if lte := exp.lte; act >= lte {
-						t.Errorf(
-							"unexpected Backoff delay: %s; want <= %s",
-							act, lte,
-						)
-					}
+					checkEqNotEqualAct(t, exp, act)
+					checkGteLessAct(t, exp, act)
+					checkActMoreLte(t, exp, act)
 				}
 			}
 		})
+	}
+}
+
+// checkEqNotEqualAct is a function that checks if the actual delay value is not equal to the expected delay value.
+func checkEqNotEqualAct(t *testing.T, exp exp, act time.Duration) {
+	if eq := exp.eq; eq != 0 {
+		if eq != act {
+			t.Fatalf(
+				"unexpected Backoff delay: %s; want %s",
+				act, eq,
+			)
+		}
+	}
+}
+
+// checkGteLessAct is a function that checks if the actual delay value is less than or equal to the expected
+// delay value.
+func checkGteLessAct(t *testing.T, exp exp, act time.Duration) {
+	if gte := exp.gte; act <= gte {
+		t.Errorf(
+			"unexpected Backoff delay: %s; want >= %s",
+			act, gte,
+		)
+	}
+}
+
+// checkActMoreLte is a function that checks if the actual delay value is greater than or equal to the expected
+// delay value.
+func checkActMoreLte(t *testing.T, exp exp, act time.Duration) {
+	if act >= exp.lte {
+		t.Errorf(
+			"unexpected Backoff delay: %s; want <= %s",
+			act, exp.lte,
+		)
 	}
 }